+++ CHANGES.rst
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 
+* Add :attr: email to :class: `oauth1.Twitter` provider.
 * Fix Twitter user info endpoint to include image url.
 * Support passing of user state in oauth2 to support variable redirect urls.
 * The :class:`.oauth2.Google` provider now uses